+++ CHANGELOG.md
@@ -8,6 +8,11 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 - Simplified building actions (`multipleProps: true|false` is no longer needed), `getActions()` now returns both multipleProp and singleProp actions (whole object setters + individual property setters) and you can just pick and choose which ones you need
 - Removed `mediaPropActions` which is redundant
 
+### Added
+- New video-iframe component.
+- New Vimeo block.
+- New Youtube block.
+
 ## [3.5.0] - 2020-06-03
 
 ### Added